refactor: split mongo indexation into files
