+++ src/imports/meta/loadMetaObjects.ts
@@ -1,14 +1,8 @@
-import BluebirdPromise from 'bluebird';
-
 import chokidar from 'chokidar';
 import fs from 'fs';
 import glob from 'glob';
 
 import debounce from 'lodash/debounce';
-import isArray from 'lodash/isArray';
-import isEmpty from 'lodash/isEmpty';
-import isNumber from 'lodash/isNumber';
-import isObject from 'lodash/isObject';
 import unset from 'lodash/unset';
 
 import { Document } from '@imports/model/Document';
@@ -18,25 +12,16 @@ import { checkInitialData } from '../data/initialData';
 import { db } from '../database';
 import { MetaAccess } from '../model/MetaAccess';
 import { logger } from '../utils/logger';
+import applyIndexes from './applyIndexes';
 import buildReferences from './buildReferences';
 
 const rebuildReferencesDelay = 1000;
-
-const dropAllIndexes = false;
 
 const rebuildReferences = debounce(function () {
 	logger.info('[kondata] Rebuilding references');
 
 	MetaObject.References = buildReferences(MetaObject.Meta);
 }, rebuildReferencesDelay);
-
-async function tryEnsureIndex({ collection, fields, options }: { collection: (typeof MetaObject.Collections)[string]; fields: { [key: string]: 1 }; options: { name: string } }) {
-	try {
-		await collection.createIndex(fields, options);
-	} catch (e) {
-		logger.trace('Index Error: ', e);
-	}
-}
 
 async function registerMeta(meta: MetaObjectType) {
 	if (meta.type !== 'composite' && meta.type !== 'document') {
@@ -66,189 +51,9 @@ async function registerMeta(meta: MetaObjectType) {
 		MetaObject.Collections[`${meta.name}.History`] = db.collection(`${meta.collection ?? meta.name}.History`);
 		MetaObject.Collections[`${meta.name}.Trash`] = db.collection(`${meta.collection ?? meta.name}.Trash`);
 		MetaObject.Collections[`${meta.name}.AutoNumber`] = db.collection(`${meta.collection ?? meta.name}.AutoNumber`);
-
-		const processIndexes = async function () {
-			// Drop all indexes of meta
-			if (dropAllIndexes) {
-				await MetaObject.Collections[meta.name].dropIndexes();
-				await MetaObject.Collections[`${meta.name}.Comment`].dropIndexes();
-				await MetaObject.Collections[`${meta.name}.History`].dropIndexes();
-			}
-
-			// Create indexes for history collections
-			const historyIndexes = ['dataId', 'createdAt'];
-			await BluebirdPromise.each(historyIndexes, async historyIndex => {
-				await tryEnsureIndex({
-					collection: MetaObject.Collections[`${meta.name}.History`],
-					fields: { [historyIndex]: 1 },
-					options: { name: historyIndex },
-				});
-			});
-
-			// Create indexes for comment collections
-			const commentIndexes = ['dataId', '_createdAt'];
-			await BluebirdPromise.each(commentIndexes, async commentIndex => {
-				await tryEnsureIndex({
-					collection: MetaObject.Collections[`${meta.name}.Comment`],
-					fields: { [commentIndex]: 1 },
-					options: { name: commentIndex },
-				});
-			});
-
-			const iterableMeta = Object.keys(meta.fields).map((key: string) => key);
-
-			await BluebirdPromise.each(iterableMeta, async (fieldName: string) => {
-				const field = meta.fields[fieldName];
-
-				if (!['richText', 'composite'].includes(field.type)) {
-					if (field.isSortable === true || field.isUnique === true || ['lookup', 'address', 'autoNumber'].includes(field.type)) {
-						let subFields = [''];
-
-						switch (field.type) {
-							case 'lookup':
-								subFields = ['._id'];
-								break;
-							case 'email':
-								subFields = ['.address'];
-								break;
-							case 'money':
-								subFields = ['.value'];
-								break;
-							case 'personName':
-								subFields = ['.full'];
-								break;
-							case 'phone':
-								subFields = ['.phoneNumber', '.countryCode'];
-								break;
-							case 'address':
-								subFields = ['.country', '.state', '.city', '.district', '.place', '.number', '.complement', '.postalCode', '.placeType'];
-								break;
-						}
-
-						const options: {
-							unique: number;
-							name: string;
-							sparse?: number;
-						} = {
-							unique: 0,
-							name: fieldName,
-						};
-
-						if (field.type === 'autoNumber' || field.isUnique === true) {
-							options.unique = 1;
-						}
-
-						if (field.isUnique === true && field.isRequired !== true) {
-							options.sparse = 1;
-						}
-
-						if (['username', 'emails'].includes(field.name) && meta.name === 'User') {
-							options.unique = 1;
-							options.sparse = 1;
-						}
-
-						const fields = subFields.reduce((acc, subField) => {
-							acc[fieldName + subField] = 1;
-							return acc;
-						}, {} as any);
-
-						logger.info(`Ensure Index at ${meta.collection}: ${fieldName}`);
-						await tryEnsureIndex({
-							collection: MetaObject.Collections[meta.name],
-							fields,
-							options,
-						});
-					}
-				}
-			});
-
-			// Create indexes for internal fields
-			const metaDefaultIndexes = ['_user._id', '_user.group._id', '_updatedAt', '_updatedBy._id', '_createdAt', '_createdBy._id'];
-
-			await BluebirdPromise.each(metaDefaultIndexes, async metaDefaultIndex =>
-				tryEnsureIndex({
-					collection: MetaObject.Collections[meta.name],
-					fields: { [metaDefaultIndex]: 1 },
-					options: { name: metaDefaultIndex },
-				}),
-			);
-
-			// Create indexes defined in meta
-			if (meta.indexes != null && isObject(meta.indexes) && !isArray(meta.indexes) && Object.keys(meta.indexes).length > 0) {
-				await BluebirdPromise.each(Object.keys(meta.indexes), async indexName => {
-					let index = meta.indexes?.[indexName];
-
-					if (index == null) {
-						index = {
-							keys: {},
-							options: {
-								name: indexName,
-							},
-						};
-					}
-
-					if (index.keys == null) {
-						index.keys = {};
-					}
-					if (index.options == null) {
-						index.options = {};
-					}
-					if (index.options.name == null) {
-						index.options.name = indexName;
-					}
-
-					logger.info(`Ensure Index at ${meta.collection}: ${index.options.name}`);
-					if (Object.keys(index.keys).length > 0) {
-						const keys = Object.entries(index.keys).reduce((acc, [key, direction]) => {
-							acc[key.replace(/:/g, '.')] = direction;
-							return acc;
-						}, {} as any);
-
-						await tryEnsureIndex({
-							collection: MetaObject.Collections[meta.name],
-							fields: keys,
-							options: { name: index.options.name },
-						});
-					}
-				});
-			}
-
-			// Create text index
-			if (meta.indexText != null && isObject(meta.indexText) && !isArray(meta.indexText) && Object.keys(meta.indexText).length > 0) {
-				const options: {
-					name: string;
-					default_language: string;
-					weights: Record<string, number>;
-				} = {
-					name: 'TextIndex',
-					default_language: MetaObject.Namespace.language,
-					weights: {},
-				};
-
-				logger.info(`Ensure Index at ${meta.collection}: ${options.name}`);
-
-				const keys = Object.entries(meta.indexText).reduce((acc, [key, weight]) => {
-					acc[key.replace(/:/g, '.')] = 'text';
-					if (isNumber(weight) && weight > 0) {
-						options.weights[key.replace(/:/g, '.')] = weight;
-					}
-					return acc;
-				}, {} as any);
-
-				await tryEnsureIndex({
-					collection: MetaObject.Collections[meta.name],
-					fields: keys,
-					options,
-				});
-			}
-		};
-
-		if (isEmpty(process.env.DISABLE_REINDEX) || process.env.DISABLE_REINDEX === 'false' || process.env.DISABLE_REINDEX === '0') {
-			processIndexes().catch(e => {
-				logger.error(e, `Error creating indexes for ${meta.name}`);
-			});
-		}
 	}
+
+	await applyIndexes(meta);
 }
 
 const deregisterMeta = function (meta: any) {
@@ -337,7 +142,6 @@ function dbWatch() {
 		rebuildReferences();
 	});
 }
-
 
 const fsLoad = (metadataDir: string) => {
 	logger.info(`Loading MetaObject.Meta from directory ${metadataDir} ...`);
